Rename the test namespace for IdGenerators to match the renamed namespace in Shared Kernel​
